fix: changes in versioning for API's and in structure.
